Cleanup tools/emprofile.py and add more testing. NFC (#13403)

+++ tools/emprofile.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 # Copyright 2016 The Emscripten Authors.  All rights reserved.
 # Emscripten is available under two separate licenses, the MIT license and the
 # University of Illinois/NCSA Open Source License.  Both these licenses can be
@@ -16,34 +16,24 @@ profiler_logs_path = os.path.join(tempfile.gettempdir(), 'emscripten_toolchain_p
 # If set to 1, always generates the output file under the same filename and doesn't delete the temp data.
 DEBUG_EMPROFILE_PY = 0
 
-OUTFILE = 'toolchain_profiler.results_' + time.strftime('%Y%m%d_%H%M')
-for arg in sys.argv:
-  if arg.startswith('--outfile='):
-    OUTFILE = arg.split('=', 1)[1].strip().replace('.html', '')
-
 
 # Deletes all previously captured log files to make room for a new clean run.
 def delete_profiler_logs():
-  try:
+  if os.path.exists(profiler_logs_path):
     shutil.rmtree(profiler_logs_path)
-  except IOError:
-    pass
 
 
 def list_files_in_directory(d):
   files = []
-  try:
-    items = os.listdir(d)
-    for i in items:
+  if os.path.exists(d):
+    for i in os.listdir(d):
       f = os.path.join(d, i)
       if os.path.isfile(f):
-        files += [f]
-    return files
-  except IOError:
-    return []
+        files.append(f)
+  return files
 
 
-def create_profiling_graph():
+def create_profiling_graph(outfile):
   log_files = [f for f in list_files_in_directory(profiler_logs_path) if 'toolchain_profiler.pid_' in f]
 
   all_results = []
@@ -63,16 +53,16 @@ def create_profiling_graph():
       print('Failed to parse JSON file "' + f + '"!', file=sys.stderr)
       sys.exit(1)
   if len(all_results) == 0:
-    print('No profiler logs were found in path "' + profiler_logs_path + '". Try setting the environment variable EM_PROFILE_TOOLCHAIN=1 and run some emcc commands, and then rerun "python emprofile.py --graph" again.')
-    return
+    print(f'No profiler logs were found in path: ${profiler_logs_path}.\nTry setting the environment variable EM_PROFILE_TOOLCHAIN=1 and run some emcc commands, then re-run "emprofile.py --graph".', file=sys.stderr)
+    return 1
 
   all_results.sort(key=lambda x: x['time'])
 
-  json_file = OUTFILE + '.json'
+  json_file = outfile + '.json'
   open(json_file, 'w').write(json.dumps(all_results, indent=2))
   print('Wrote "' + json_file + '"')
 
-  html_file = OUTFILE + '.html'
+  html_file = outfile + '.html'
   html_contents = open(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'toolchain_profiler.results_template.html'), 'r').read().replace('{{{results_log_file}}}', '"' + json_file + '"')
   open(html_file, 'w').write(html_contents)
   print('Wrote "' + html_file + '"')
@@ -80,9 +70,13 @@ def create_profiling_graph():
   if not DEBUG_EMPROFILE_PY:
     delete_profiler_logs()
 
+  return 0
 
-if len(sys.argv) < 2:
-  print('''Usage:
+
+def main(args):
+  if len(args) < 2:
+    print('''\
+Usage:
        emprofile.py --reset
          Deletes all previously recorded profiling log files.
 
@@ -93,14 +87,23 @@ Optional parameters:
 
         --outfile=x.html
           Specifies the name of the results file to generate.
-''')
-  sys.exit(1)
+  ''')
+    return 1
+
+  if '--reset' in args:
+    delete_profiler_logs()
+  elif '--graph' in args:
+    outfile = 'toolchain_profiler.results_' + time.strftime('%Y%m%d_%H%M')
+    for arg in args:
+      if arg.startswith('--outfile='):
+        outfile = arg.split('=', 1)[1].strip().replace('.html', '')
+    return create_profiling_graph(outfile)
+  else:
+    print('Unknown command "' + args[1] + '"!')
+    return 1
+
+  return 0
 
 
-if '--reset' in sys.argv:
-  delete_profiler_logs()
-elif '--graph' in sys.argv:
-  create_profiling_graph()
-else:
-  print('Unknown command "' + sys.argv[1] + '"!')
-  sys.exit(1)
+if __name__ == '__main__':
+  sys.exit(main(sys.argv))
